5467 enforce email verification
